BACK-2777: Port all existing commands to the flex namespace
