+++ lib/commands/flex/config.js
@@ -14,25 +14,31 @@
  */
 
 const async = require('async');
-const service = require('../service.js');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const init = require('../../init.js');
+const project = require('../../project.js');
+const user = require('../../user.js');
+const util = require('../../util.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function status(argv, cb) {
+function initUrl(host, cb) {
+  if (host != null) user.host = util.formatHost(host);
+  cb();
+}
+
+function configure(argv, cb) {
   init(argv);
   return async.series([
+    (next) => initUrl(argv.instance, next),
     (next) => user.setup(argv, next),
-    (next) => project.restore(next),
-    (next) => service.serviceStatus(next)
+    (next) => project.config(argv, next),
+    (next) => user.save(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'status',
-  desc: 'return the health of a Flex Service cluster',
-  handler: status
+  command: 'config [instance]',
+  desc: "set project options (including optional Kinvey instance, e.g. 'acme-us1')",
+  handler: configure
 };
