+++ lib/commands/flex/deploy.js
@@ -14,21 +14,27 @@
  */
 
 const async = require('async');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const config = require('config');
+const service = require('../../service.js');
+const init = require('../../init.js');
+const project = require('../../project.js');
+const user = require('../../user.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function logout(argv, cb) {
-  return async.series([
-    (next) => user.logout(next),
-    (next) => project.logout(next)
+function deploy(argv, cb) {
+  init(argv);
+  return async.waterfall([
+    (next) => user.setup(argv, next),
+    (next) => project.restore(next),
+    (next) => service.validate(config.paths.package, next),
+    (version, next) => service.deploy(config.paths.package, version, next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'logout',
-  desc: 'resets host, removes sessions, and clears project settings',
-  handler: logout
+  command: 'deploy',
+  desc: 'deploy the current project to the Kinvey FlexService Runtime',
+  handler: deploy
 };
