+++ lib/commands/flex/index.js
@@ -13,27 +13,16 @@
  * contents is a violation of applicable laws.
  */
 
-const async = require('async');
-const service = require('../service.js');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
-
-function status(argv, cb) {
-  init(argv);
-  return async.series([
-    (next) => user.setup(argv, next),
-    (next) => project.restore(next),
-    (next) => service.jobStatus(argv.id, next)
-  ], (err) => {
-    // TODO: cb should receive result from tasks
-    handleActionFailure(err, cb);
-  });
-}
-
-module.exports = {
-  command: 'job [id]',
-  desc: 'return the job status of a deploy command',
-  handler: status
+module.exports = yargs => {
+  return yargs
+    .usage('kinvey flex <command> [args] [options]')
+    .command(require('./config'))
+    .command(require('./deploy'))
+    .command(require('./job'))
+    .command(require('./list'))
+    .command(require('./logout'))
+    .command(require('./logs'))
+    .command(require('./recycle'))
+    .command(require('./status'))
+    .demand(1, '');
 };
