+++ lib/commands/flex/job.js
@@ -14,25 +14,26 @@
  */
 
 const async = require('async');
-const service = require('../service.js');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const service = require('../../service.js');
+const init = require('../../init.js');
+const project = require('../../project.js');
+const user = require('../../user.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function recycle(argv, cb) {
+function status(argv, cb) {
   init(argv);
   return async.series([
     (next) => user.setup(argv, next),
     (next) => project.restore(next),
-    (next) => service.recycle(next)
+    (next) => service.jobStatus(argv.id, next)
   ], (err) => {
+    // TODO: cb should receive result from tasks
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'recycle',
-  desc: 'recycle the Service',
-  handler: recycle
+  command: 'job [id]',
+  desc: 'return the job status of a deploy command',
+  handler: status
 };
