+++ lib/commands/flex/list.js
@@ -14,31 +14,24 @@
  */
 
 const async = require('async');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const util = require('../util.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const init = require('../../init.js');
+const project = require('../../project.js');
+const user = require('../../user.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function initUrl(host, cb) {
-  if (host != null) user.host = util.formatHost(host);
-  cb();
-}
-
-function configure(argv, cb) {
+function list(argv, cb) {
   init(argv);
   return async.series([
-    (next) => initUrl(argv.instance, next),
     (next) => user.setup(argv, next),
-    (next) => project.config(argv, next),
-    (next) => user.save(next)
+    (next) => project.restore(next),
+    (next) => project.list(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'config [instance]',
-  desc: "set project options (including optional Kinvey instance, e.g. 'acme-us1')",
-  handler: configure
+  command: 'list',
+  desc: 'list Internal Flex Services for the current app',
+  handler: list
 };
