+++ lib/commands/flex/logout.js
@@ -14,24 +14,21 @@
  */
 
 const async = require('async');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const project = require('../../project.js');
+const user = require('../../user.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function list(argv, cb) {
-  init(argv);
+function logout(argv, cb) {
   return async.series([
-    (next) => user.setup(argv, next),
-    (next) => project.restore(next),
-    (next) => project.list(next)
+    (next) => user.logout(next),
+    (next) => project.logout(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'list',
-  desc: 'list Internal Flex Services for the current app',
-  handler: list
+  command: 'logout',
+  desc: 'resets host, removes sessions, and clears project settings',
+  handler: logout
 };
