+++ lib/commands/flex/recycle.js
@@ -14,27 +14,25 @@
  */
 
 const async = require('async');
-const config = require('config');
-const service = require('../service.js');
-const init = require('../init.js');
-const project = require('../project.js');
-const user = require('../user.js');
-const handleActionFailure = require('../util').handleCommandFailure;
+const service = require('../../service.js');
+const init = require('../../init.js');
+const project = require('../../project.js');
+const user = require('../../user.js');
+const handleActionFailure = require('../../util').handleCommandFailure;
 
-function deploy(argv, cb) {
+function recycle(argv, cb) {
   init(argv);
-  return async.waterfall([
+  return async.series([
     (next) => user.setup(argv, next),
     (next) => project.restore(next),
-    (next) => service.validate(config.paths.package, next),
-    (version, next) => service.deploy(config.paths.package, version, next)
+    (next) => service.recycle(next)
   ], (err) => {
     handleActionFailure(err, cb);
   });
 }
 
 module.exports = {
-  command: 'deploy',
-  desc: 'deploy the current project to the Kinvey FlexService Runtime',
-  handler: deploy
+  command: 'recycle',
+  desc: 'recycle the Service',
+  handler: recycle
 };
