+++ test/integration/job.test.js
@@ -34,7 +34,7 @@ function assertLoggerForStatus(spy) {
 describe('job', () => {
   const mockServer = new MockServer(true);
   const sandbox = sinon.sandbox.create();
-  const cmdJobPath = './../../lib/commands/job';
+  const cmdJobPath = './../../lib/commands/flex/job';
 
   afterEach('generalCleanup', (cb) => {
     sandbox.restore();
