+++ test/integration/status.test.js
@@ -27,7 +27,7 @@ describe('status', () => {
   const mockServer = new MockServer(true);
   const sandbox = sinon.sandbox.create();
 
-  const cmdStatusPath = './../../lib/commands/status';
+  const cmdStatusPath = './../../lib/commands/flex/status';
 
   afterEach('generalCleanup', (cb) => {
     sandbox.restore();
