+++ test/unit/commands/logs.test.js
@@ -16,7 +16,7 @@
 const yargs = require('yargs');
 const sinon = require('sinon');
 const service = require('./../../../lib/service.js');
-const logs = require('../../../lib/commands/logs.js').handler;
+const logs = require('../../../lib/commands/flex/logs.js').handler;
 const pkg = require('./../../../package.json');
 const project = require('./../../../lib/project.js');
 const user = require('./../../../lib/user.js');
